[bphh-935] Refactor logic to only show jurisdiction setup call out warning banner only if there isn't at least one valid contact per permit type
