+++ app/frontend/components/domains/jurisdictions/submission-inbox/jurisdiction-submisson-inbox-screen.tsx
@@ -5,6 +5,7 @@ import { observer } from "mobx-react-lite"
 import React from "react"
 import { useTranslation } from "react-i18next"
 import { useJurisdiction } from "../../../../hooks/resources/use-jurisdiction"
+import { usePermitClassificationsLoad } from "../../../../hooks/resources/use-permit-classifications-load"
 import { useSearch } from "../../../../hooks/use-search"
 import { IPermitApplication } from "../../../../models/permit-application"
 import { useMst } from "../../../../setup/root"
@@ -27,6 +28,7 @@ export const JurisdictionSubmissionInboxScreen = observer(function JurisdictionS
   const { t } = useTranslation()
   const { permitApplicationStore } = useMst()
   const { currentJurisdiction, error } = useJurisdiction()
+  const { isLoaded: isPermitClassificationsLoaded } = usePermitClassificationsLoad()
 
   const { currentPage, totalPages, totalCount, countPerPage, handleCountPerPageChange, handlePageChange, isSearching } =
     permitApplicationStore
@@ -34,7 +36,7 @@ export const JurisdictionSubmissionInboxScreen = observer(function JurisdictionS
   useSearch(permitApplicationStore, [currentJurisdiction?.id])
 
   if (error) return <ErrorScreen error={error} />
-  if (!currentJurisdiction) return <LoadingScreen />
+  if (!currentJurisdiction || !isPermitClassificationsLoaded) return <LoadingScreen />
 
   return (
     <Container maxW="container.lg" p={8} as={"main"}>
